requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-Migrate==3.1.0
Flask-WTF
Flask-Bootstrap
Flask-Reuploaded
Jinja2==3.1.2
Werkzeug==2.2.2
python-dotenv
python-decouple
requests
mysqlclient
